Fix time_limit_sec for GLOP and PDLP (#1859)
Fixes #1858

+++ cvxpy/reductions/solvers/conic_solvers/glop_conif.py
@@ -152,7 +152,7 @@ class GLOP(ConicSolver):
             if not solver.SetSolverSpecificParametersAsString(proto_str):
                 return {"status": s.SOLVER_ERROR}
         if "time_limit_sec" in solver_opts:
-            solver.SetTimeLimit(1000 * float(solver_opts["time_limit_sec"]))
+            solver.SetTimeLimit(int(1000 * solver_opts["time_limit_sec"]))
         solver.Solve()
         solver.FillSolutionResponseProto(response)
